VRV-6362 Fix for building not rendering for a single frame when disabling SFX sensor
Sensor FX switches between two different rendering modes when toggling sensors. The osgEarthBuildings BuildingsPager has functionality which is tracking if geometry was rendered last frame but none this frame. When that happens it invokes releaseGlObjects. The switching of rendering modes would cause this to happen every time as one mode does not register any render leaves.